// src/interfaces/PoseBodyInfo.ts
export default interface PoseBodyInfo {
    bodyCenter: { x: number, y: number };
    groundPosY: number;
    kneePosY: number;
    hipPosY: number;
    rightShoulderPosX: number;
    leftShoulderPosX: number;
    rightShoulderPosXOffset: number;
    leftShoulderPosXOffset: number;
    shoulderPosY: number;
}